requirements.txt
streamlit
sentence-transformers
scikit-learn
Pillow
numpy
